Codechange: Move GrfProcessingState to separate file.
Allows reuse by other files.
