+++ src/newgrf/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_files(
     newgrf_bytereader.cpp
     newgrf_bytereader.h
+    newgrf_internal.h
 )
